Incorporated some review comments.
