+++ tests/mock.py
@@ -77,13 +77,25 @@ class JWPlatformMock(HttpApiMock):
             response=lambda _: (200,
                                 {"upload_id": "uploADid",
                                  "upload_token": "upload_token",
-                                 "upload_link": "http://s3server/upload-link",
-                                 "upload_method": "upload_method"}
+                                 "upload_link": "http://s3server/upload-link"
+                                 }
                                 )
         ),
         HttpApiMockEndpoint(
             operation_id='completeUpload',
             match_pattern=b'^PUT /v1/uploads/(?P<upload_id>[A-Za-z0-9]{8}?)/complete',
             response=lambda _: (202, {})
+        )
+    ]
+
+
+class S3Mock(HttpApiMock):
+    hostnames = ['s3server']
+
+    endpoints = [
+        HttpApiMockEndpoint(
+            operation_id='uploadToS3',
+            match_pattern=b'^PUT ',
+            response=lambda _: (200, None),
         )
     ]
